// src/app/models/course.model.ts
export default class Course {
	club_id?: string;
	club_name?:string;
	club_published?: boolean;
	club_description?: string;
	club_image?: string;
	club_membership?: string;
	number_of_holes?: number;
	address?: string;
	city?: string;
	state?: string;
	holes?: number;
	par?: number;
	slope?: number;
	rating?: number;
	country?: string;
	postal_code?: string;
	phone?: string;
	fax?: string;
	website?: string;
	contact_name?: string;
	contact_title?: string;
	email_address?: string;
	driving_range?: boolean;
	putting_green?: boolean;
	chipping_green?: boolean;
	practice_bunker?: boolean;
	motor_cart?: boolean;
	pull_cart?: boolean;
	golf_clubs_rental?: boolean;
	club_fitting?: boolean;
	pro_shop?: boolean;
	golf_lessons?: boolean;
	caddie_hire?: boolean;
	restaurant?: boolean;
	reception_hall?: boolean;
	changing_room?: boolean;
	lockers?: boolean;
	lodging_on_site?: boolean;


}